+++ tsconfig.json
@@ -39,6 +39,6 @@
 
     "types": ["kakao.maps.d.ts"]
   },
-  "include": ["src", "custom.d.ts", "firebase-messaging-sw.ts"],
+  "include": ["src", "custom.d.ts", "firebase-messaging-sw.js"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
